[feat] #34 StoreRepository StoreModel 을 반환할 수 있도록 변경
